Add identity test for m_inverse
This test fails (see testResults.txt) It should output the identity matrix
since the inverse of the Identity is the Identity. It outputs a matrix
with three's in the bottom four rows. This is not expected behavior.

A possible fix would be to take the norm of the matrix, but this may not
fix all the associated problem, only this one case.


Former-commit-id: 93eef7c8c57bfb942dd62a571717c3ec64daecb3