Updated the group 4 test matrix.
tested the three functions of group four to see if they work for relatively
simple and straightforward testing. Used a 6x6 matrix and a 2x4 matrix to
test the functions and all went according to plan, so these functions should
be fine. Possibly not work for more complicated test cases, but as of now that
is uneccessary to test.


Former-commit-id: b3d2c6ca073b952658a2b2bec91b63bc74149f0b