Begin to add blas functionality to invers function
This is the same process I would like the shared team to use in order
to begin using blas. If you have any questions, please email me or
get on the groupme and ask. The jist is that we will be using the blas
functions directly in the code, adn you simply go to the documentation
and try to match parameters to each function.

This is a crucial step for you to document. The documentation I have in
these changes is the BARE MINIMUM for you to use. If you do not
document this step, then any problems we have in the future will require
use to go back to the drawing board for that function.

This particular commit includes part of the implementation of these
blas functions, and will be updated in the future.


Former-commit-id: 39f584587245d433463edb99ef41ed7b48142484