Adding new ICA Matlab files
Adding files from Helsinki University for Matlab implementation of ICA
algorithm. files are as follows:
fastica  -> implementation of ICA
fpica    -> underlying algorithm and bulk of code
pcamat   -> pca implementation
remmean  -> return the matrix with mean taken out
whitenv  -> whiten the matrix or 'spherex' it
createdb -> set up the matrix to be analyzed
rec      -> recognize faces based on output for ICA

These are the pieces of the puzzle, next step is to put it all together.